requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
certifi==2024.2.2
decorator==5.1.1
h11==0.14.0
idna==3.6
iniconfig==2.0.0
loguru==0.7.2
outcome==1.3.0.post0
packaging==23.2
pluggy==1.4.0
py==1.11.0
PySocks==1.7.1
pytest==8.0.0
retry==0.9.2
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.0
wsproto==1.2.0
